- more work to make cseq unsigned
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@6416 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ main/resip/dum/Dialog.hxx
@@ -132,8 +132,8 @@ class Dialog
       
       //variants
       NameAddr mLocalContact;
-      unsigned long mLocalCSeq;
-      unsigned long mRemoteCSeq;
+      unsigned int mLocalCSeq;
+      unsigned int mRemoteCSeq;
       NameAddr mRemoteTarget;
       NameAddr mLocalNameAddr;
       NameAddr mRemoteNameAddr;
